requirements.txt
accelerate==1.7.0
bitsandbytes==0.48.2
click==8.1.7
datasets==3.5.0
diffusers==0.32.2
einops==0.8.1
numpy==2.3.5
open_clip_torch==2.24.0
packaging==25.0
Pillow==12.0.0
pyspng==0.1.3
pytorch_lightning==2.2.5
Requests==2.32.5
scipy==1.16.3
torch==2.2.2
torch_xla==2.9.0
torchvision==0.17.2
tqdm==4.66.4
transformers==4.51.3
xformers==0.0.25.post1
